Upgrade non dask

+++ python/hidebound/core/traits_test.py
@@ -12,7 +12,7 @@ import hidebound.core.traits as traits
 class TraitsTests(unittest.TestCase):
     def test_get_image_width(self):
         with TemporaryDirectory() as root:
-            img = np.zeros((2, 4, 3))
+            img = np.zeros((2, 4, 3), dtype=np.uint8)
             filepath = Path(root, 'foo.png')
             skimage.io.imsave(filepath.as_posix(), img)
             result = traits.get_image_width(filepath)
@@ -20,7 +20,7 @@ class TraitsTests(unittest.TestCase):
 
     def test_get_image_height(self):
         with TemporaryDirectory() as root:
-            img = np.zeros((2, 4, 3))
+            img = np.zeros((2, 4, 3), dtype=np.uint8)
             filepath = Path(root, 'foo.png')
             skimage.io.imsave(filepath.as_posix(), img)
             result = traits.get_image_height(filepath)
@@ -28,13 +28,13 @@ class TraitsTests(unittest.TestCase):
 
     def test_get_num_image_channels(self):
         with TemporaryDirectory() as root:
-            img = np.zeros((2, 4, 3))
+            img = np.zeros((2, 4, 3), dtype=np.uint8)
             filepath = Path(root, 'foo.png')
             skimage.io.imsave(filepath.as_posix(), img)
             result = traits.get_num_image_channels(filepath)
             self.assertEqual(result, 3)
 
-            img = np.zeros((2, 4))
+            img = np.zeros((2, 4), dtype=np.uint8)
             filepath = Path(root, 'bar.jpg')
             skimage.io.imsave(filepath.as_posix(), img)
             result = traits.get_num_image_channels(filepath)
